requirements.txt
nltk==3.8.1
newspaper3k==0.2.8
gtts==2.2.3
ttkthemes==3.2.2
pydub==0.25.1